出场

+++ src/main/java/com/xxm/parking/pojo/Plot.java
@@ -20,7 +20,7 @@ public class Plot {
 	private int tel;
 	private int totalparkinglot;
 	private int leftparkinglot;
-	private BigDecimal charge;//收费  元/分钟
+	private double charge;//收费  元/分钟
 	private String description;
 	private String map;//地图存放地址
 	private int freetime;//免费时长 分钟
